[Feat] RenderManager loading Scenes
RenderManager loading Scenes, IE removing hard coding in the rendering code. The core concept of scenes has been implemented. Allowing for all data to be sent to the renderer from outside the renderer.  Some more features are needed to fully support scenes:
XML or JSON scene saver & interpreter.
For fragment shaders, texture samplers can be access via different bindings. This means that every model with a unique texture needs a unique fragment shader. To fix this, I will send all in use textureSamplers to the fragment shader and send the Index in a push constant. This might mean there is a maximum amount of samplers allowed, and if the number of samplers must be a hard number, IE 16. The buffer must either work without all samplers being actively bound or having the samplers not yet bound be filled with a default sampler. Another potential solution would be to have different descriptor sets per object. This has the advantage of working on all hardware, but comes with a potential hit on performance. I must investigate further.