[Fix] Image Sampler Indexing
Previously, Image Samplers in the frag shader was being indexed by a value send from the vert shader. This does not work because the index is required to be a uniform throughout each fragment shader call. The fix was to send the push constant holding the index straight to the fragment shader.